requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
supabase-py==0.12.0
pydantic==2.7.4
pydantic-settings==2.3.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.3
pandas==2.2.2
requests==2.32.3
